* renamed binding outer directory for nicer path * moved example file from ffi/main.go -> ./example.go
