+++ sdk/ffi/go/build.sh
@@ -5,8 +5,7 @@ PROJECT_NAME="go"
 set -eu
 MODE="--release"
 
-GO_DIR="./ffi"
-GO_PATH="${GO_DIR}/bindings"
+GO_DIR="./go-nym"
 UDL_PATH="./src/bindings.udl"
 
 build_artifacts() {
@@ -18,20 +17,17 @@ build_artifacts() {
   printf "bindings built \n\n"
 
   # something not right with these - having to add it manually to bindings.go for the moment
-  #pushd $GO_DIR/bindings
-  #echo $(pwd)
-  #LD_LIBRARY_PATH="${LD_LIBRARY_PATH:-}:../../target/release" \
+  # pushd $GO_DIR/bindings
+  # echo $(pwd)
+  # LD_LIBRARY_PATH="${LD_LIBRARY_PATH:-}:../../target/release" \
   #	CGO_LDFLAGS="-L../target/release -lnym_go_ffi -lm -ldl" \
   #	CGO_ENABLED=1 \
   #	go run ../main.go
 }
 
 clean_artifacts() {
-  # the path to the new folder we are including
-  GO_DIR="./ffi"
-  GO_PATH="${GO_DIR}/bindings"
   # clean up existing things
-  rm -rf $GO_PATH
+  rm -rf $GO_DIR
   cargo clean
 }
